Move util functions from ledger.go to utils.go
